pyproject.toml
[tool.poetry]
name = "pynarrator"
version = "0.1.0"
description = ""
authors = ["neumann-mlucas <neumann.m.lucas@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
pyinstaller = "^6.4.0"
pygame = "^2.5.2"
deep-translator = "^1.11.4"
pdbp = "^1.5.0"
pillow = "^10.2.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
